@@ -1,27 +1,34 @@
 import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
+import { NgModule, LOCALE_ID } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { HttpModule } from '@angular/http';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { MatButtonModule, MatCheckboxModule } from '@angular/material';
+import { HttpClientModule } from '@angular/common/http';
+
+import { TextMaskModule } from 'angular2-text-mask';
 
 import { HomeComponent } from './pages/home/home.component';
-import { EmployeesListComponent } from './components/employees-list/employees-list.component';
+import { EmployeesListComponent } from './pages/home/components/employees-list/employees-list.component';
+import { ModalComponent } from './pages/home/components/modal/modal.component';
+import { EmployeeFormComponent } from './pages/home/components/employee-form/employee-form.component';
+import { NotificationComponent } from './components/notification/notification.component';
+import { ConfirmComponent } from './components/confirm/confirm.component';
+import { MultiSelectComponent } from './components/multi-select/multi-select.component';
 
 @NgModule({
   declarations: [
     HomeComponent,
-    EmployeesListComponent
+    EmployeesListComponent,
+    ModalComponent,
+    EmployeeFormComponent,
+    NotificationComponent,
+    ConfirmComponent,
+    MultiSelectComponent
   ],
   imports: [
     BrowserModule,
     FormsModule,
-    HttpModule,
-    BrowserAnimationsModule,
-    MatButtonModule,
-    MatCheckboxModule 
+    HttpClientModule,
+    TextMaskModule
   ],
-  providers: [],
-  bootstrap: [HomeComponent]
+  bootstrap: [HomeComponent],
 })
 export class AppModule { }
