@@ -1,4 +1,9 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Input, OnInit } from '@angular/core';
+import * as dayjs from 'dayjs';
+import 'dayjs/locale/pt-br';
+
+import { Employee } from 'app/interface/Employee';
+import { TypeNotification } from 'app/components/notification/notification.component';
 
 @Component({
   selector: 'app-home',
@@ -7,10 +12,59 @@ import { Component, OnInit } from '@angular/core';
 })
 export class HomeComponent implements OnInit{
 
-  constructor() { }
+  @Input() employeeList: Employee[];
+ 
+  isModalOpen: boolean;
 
-  ngOnInit(): void {
-    
+  type: TypeNotification;
+  message: string;
+  notificationVisible: boolean;
+
+  constructor() { 
+    this.isModalOpen = false;
+    this.employeeList = [];
+
+    this.notificationVisible = false;
   }
-  
+
+  ngOnInit() {
+    this.updateEmployeeList();
+  }
+
+  openModal() {
+    this.isModalOpen = true;
+  }
+
+  closeModal() {
+    this.isModalOpen = false;
+  }
+
+  successNotification() {
+    this.type = 'success';
+    this.message = 'Concluído';
+    this.notificationVisible = true;
+    setTimeout(() => this.notificationVisible = false, 3000);
+  }
+
+  async updateEmployeeList() {
+    
+    const response = await fetch('http://localhost:8080/employees', {
+      method: 'GET',
+      headers: {
+        'Content-Type': 'application/json',
+      }
+    });
+
+    const result = await response.json();
+
+    result.forEach(employee => {
+      employee.hireDate = dayjs(new Date(employee.hireDate)).format('DD/MM/YYYY');
+      const salaryNummber = parseFloat(employee.salary);
+      employee.salary = salaryNummber.toLocaleString('pt-BR', { style: 'currency', currency: 'BRL' });
+    });
+
+    result.sort((itemA: Employee, itemB: Employee) => itemA.name.localeCompare(itemB.name));
+    
+    this.employeeList = result;
+  }
 }
